Remove ClientInterface
No longer fits our vision
